move types to libs referenced in libs and api except CloningFilesRepositoryModule
