+++ api/libs/cloning/src/infrastructure/clone-piece-data/project-custom-features.ts
@@ -1,8 +1,16 @@
-import { JobStatus } from '@marxan-api/modules/scenarios/scenario.api.entity';
 import { FeatureTag } from '../../../../features/src';
 import { GeometrySource } from '../../../../geofeatures/src';
 
 export const projectCustomFeaturesRelativePath = 'custom-features.json';
+
+//TODO Add debt
+enum JobStatus {
+  draft = 'draft',
+  created = 'created',
+  running = 'running',
+  done = 'done',
+  failure = 'failure',
+}
 
 type FeatureData = {
   the_geom: string;
