+++ api/libs/scenario-metadata/src/index.ts
@@ -1,0 +1,4 @@
+export type ScenarioMetadata = {
+  marxanInputParameterFile?: Record<string, string | number>;
+  scenarioEditingMetadata?: Record<string, unknown>;
+};
