+++ api/libs/specification/src/feature-subset.ts
@@ -1,0 +1,6 @@
+export interface FeatureSubSet {
+  value: string;
+  target?: number;
+  fpf?: number;
+  prop?: number;
+}
