+++ api/libs/specification/src/index.ts
@@ -1,0 +1,3 @@
+export { SpecificationOperation } from './specification-operation';
+export { SpecificationFeature } from './specification-feature';
+export { FeatureSubSet } from './feature-subset';
