+++ api/libs/specification/src/specification-operation.ts
@@ -1,0 +1,5 @@
+export enum SpecificationOperation {
+  Split = 'split',
+  Stratification = 'stratification',
+  Copy = 'copy',
+}
