+++ api/tsconfig.json
@@ -49,6 +49,8 @@
       ],
       "@marxan/scenario-puvspr": ["libs/scenario-puvspr/src"],
       "@marxan/scenario-puvspr/*": ["libs/scenario-puvspr/src/*"],
+      "@marxan/scenario-metadata": ["libs/scenario-metadata/src"],
+      "@marxan/scenario-metadata/*": ["libs/scenario-metadata/src/*"],
       "@marxan/scenario-run-queue": ["libs/scenario-run-queue/src"],
       "@marxan/scenario-run-queue/*": ["libs/scenario-run-queue/src/*"],
       "@marxan/features": ["libs/features/src"],
@@ -65,6 +67,8 @@
       "@marxan/planning-units-grid/*": ["libs/planning-units-grid/src/*"],
       "@marxan/protected-areas": ["libs/protected-areas/src"],
       "@marxan/protected-areas/*": ["libs/protected-areas/src/*"],
+      "@marxan/specification": ["libs/specification/src"],
+      "@marxan/specification/*": ["libs/specification/src/*"],
       "@marxan/tiles": ["libs/tiles/src"],
       "@marxan/tiles/*": ["libs/tiles/src/*"],
       "@marxan/planning-unit-features": ["libs/planning-unit-features/src"],
